TOT-35 :  [REFACTOR] 클래스 위치 디렉토리에 맞게 수정
