+++ backend/src/main/java/com/triportreat/backend/place/domain/PlaceSearchCondition.java
@@ -1,4 +1,4 @@
-package com.triportreat.backend.place.dto;
+package com.triportreat.backend.place.domain;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
