+++ backend/src/main/java/com/triportreat/backend/place/repository/PlaceRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.triportreat.backend.place.repository;
 
-import com.triportreat.backend.place.dto.PlaceByRegionIdDto;
-import com.triportreat.backend.place.dto.PlaceSearchCondition;
+import com.triportreat.backend.place.domain.PlaceByRegionIdDto;
+import com.triportreat.backend.place.domain.PlaceSearchCondition;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
